refactor: extract userMsgIndexUtils for test coverage
- Move extractPlainText and truncateUserMsg to userMsgIndexUtils.ts
- Update ChatMessageList.vue to import from utils
- Update test file to import from utils
- This shifts coverage from component (36.4%) to util file (100%)
